requirements.txt
requests==2.7.0
ckanapi==3.6
pyOpenSSL==16.2.0
ndg-httpsclient==0.4.1
pyasn1==0.1.9
